pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.66.2",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "werkzeug>=3.1.3",
]
